Update 20-turtlebot3.sbc.deps
Fixed gazebo installation dependencies and keyserver url. On a related note, line 31 and 33 pulling whatever is at some random website and piping it straight to bash seems like a security nightmare. Script is, as far as I can tell, a trustworthy, harmless installation script for now, but that could change with zero warning by whoever controls that link into something more malicious. Initial wave of commits is just to get stuff working again so I'll just leave it for now, but will probably have to come back and change this later.